+++ CHANGELOG.md
@@ -32,6 +32,9 @@
 
 ### FEATURES/ENHANCEMENTS:
 
+* Appsec
+  * (Beta) Added URL Evasion Defense API.
+
 * DataStream
   * Added support for managing and listing `AnswerX` streams.
   * Added the [`ListAnswerXServiceIDs`](https://techdocs.akamai.com/datastream2/reference/get-answerx-ssids/) method to list `AnswerX` service IDs.